updated readme title
Since this is a hypothetial project, I thought it prudent to not use the name of an actual local business entity in the title of the readme. I think it's fine to do so in the slides, as we do, because anyone reading that far in should know what this is, but I wouldn't want this page showing up when someone tries to do a web search for that company, for instance.